Updating to 0.0.2.40
Integration with Scikit-Learn's text-analysis pipeline led the creation of the
`CorpusFromScikit` and `TermDocMatrixFromScikit` classes.

The `AutoTermSelector` class to automatically suggest terms to appear in the visualization.
This can make it easier to show large data sets, and remove fiddling with the various
minimum term frequency parameters.

For an example of how to use `CorpusFromScikit` and `AutoTermSelector`, please see demo_sklearn.py.

Also, I updated the library and examples to be compatible with spaCy 2.